arrumando dados do servidor

+++ src/index.py
@@ -1,8 +1,11 @@
 from fastapi import FastAPI, UploadFile, File, HTTPException
+from fastapi.staticfiles import StaticFiles
 from pydantic import BaseModel, ValidationError, Field
 import shutil
 import os
 from typing import List
+import math
+import fitz
 
 from service import Service
 
@@ -20,7 +23,15 @@ class Response(BaseModel):
 
 class ResponseUpload(BaseModel):
     filename : str
-    msg : str      
+    msg : str
+    # None or str
+    size : None or str
+    image_url :None or str 
+
+class ResponseUploadImage(BaseModel):
+    filename : str
+    msg : str
+    size : str              
 
 class Document(BaseModel):
     content: str
@@ -33,19 +44,79 @@ class TopResponse(BaseModel):
 
 class ReponseDocument(BaseModel):
     data: List[TopResponse]
-    query: str      
+    query: str     
+
+class ResponseStatusCode(BaseModel):
+    msg : str
+    status_code : int
+
+class ResponseUploadError(BaseModel):
+    msg : str
+
+
+app.mount("/images", StaticFiles(directory="../images"), name="images")
 
 #upload de arquivos pdf
-@app.post("/uploadfile/", status_code=201, response_model=ResponseUpload)
+@app.post("/uploadfile/", status_code=201, response_model=ResponseUpload, responses={400 : {'model' : ResponseUploadError}})
 async def create_upload_file(file: UploadFile = File(...)):
+    
+    # Verificar o tamanho do arquivo
+    file_content = await file.read()
+
+    if len(file_content) > 1000 * 1024:
+        # tratar esse erro dentro do sistema
+        raise HTTPException(status_code=400, detail="O tamanho do arquivo não pode ser maior que 500KB")
+        
     # Salvar o arquivo na pasta data com caminho absoluto
     current_dir = os.path.dirname(os.path.abspath(__file__))
     file_path = os.path.join(current_dir, '../document', file.filename)
-    
+
+    #verificar se o arquivo existe
+    if os.path.exists(file_path): 
+        file_image_name = os.path.splitext(file.filename)[0]
+        size = math.ceil(len(file_content) / 1000)
+        return {"filename": file.filename, "msg" : "Upload feito com sucesso", "size" : f"{size}KB", "image_url" : f"/images/{file_image_name}.png"}  # Adicionamos o índice 0 ao nome do arquivo
+
     with open(file_path, 'wb') as buffer:
-        shutil.copyfileobj(file.file, buffer)
+        buffer.write(file_content)
+        
+    doc = fitz.open(file_path)  
     
-    return {"filename": file.filename, "msg" : "upload realizado com sucesso"}
+    image_url = None
+    try:
+        # Percorrer cada página e obter a primeira imagem
+        for i in range(len(doc)):
+            page = doc[i]
+            pix = page.get_pixmap()
+            base = os.path.splitext(file.filename)[0]
+
+            # Ajustar o caminho para salvar as imagens
+            images_dir = os.path.join(current_dir, '../images')
+            if not os.path.exists(images_dir):
+                os.makedirs(images_dir)
+            image_path = os.path.join(images_dir, f"{base}.png")
+            pix.save(image_path)
+            image_url = f"/images/{base}.png"
+            break    
+    except Exception as e:
+        print(f'Error ao gerar o arquivo {image_url}')
+            
+    # Calcular o tamanho do arquivo
+    file_size = math.ceil(len(file_content) / 1000)
+
+    if image_url is None:
+        return {"filename": file.filename, "msg" : "upload realizado com sucesso", "size" : f"{file_size}KB", "image_url" : f"/images/gatinho.jpeg"}    
+
+    return {"filename": file.filename, "msg" : "upload realizado com sucesso", "size" : f"{file_size}KB", "image_url" : image_url}
+   
+
+#retorno do status code
+@app.get('/status', status_code=200, response_model=ResponseStatusCode, responses={500 : {'model' : ResponseStatusCode}})
+async def status(): 
+    try:
+        return {"msg" : "API em execução", "status_code": 200}
+    except Exception as e:
+        raise HTTPException(status_code=500, detail="Servidor offline")  
 
 
 #generate document response with metadata and content_text
